refactor: provider-neutral screening naming
The screening vendor is an upstream implementation detail of the
data-v3 compliance API — swapping it must not touch this app.
services/trm.ts -> services/screening.ts; docs updated.
